Basic functionality for simlating multiple touches
In order for a multitouch command to register, the pen needs to be
away from the device for roughly 10 ms or more.  We write a special
key event (0x140; couldn't find it in input-event-codes.h) that the
device uses to note this, and then grab the pen input device to keep
other inputs from coming in.  We wait a bit and then simulate
touching and releasing with touch device.  Then we lower the pen
back towards to the device, to restore the correct state.

For some reason, the first simulated multitouch tap doesn't register.
This resets when the pen is moved away from the device.  More
investigation is needed.

Also weird: If you want the input device, the touch position values
don't show up all the time.  They seem to show up on the first one,
but not later ones, until an actual multitouch event happens and the
system resets.  Perhaps because we're using the exact same touch points
in the fake events, there something that's caching the old values?

At the moment, the double touch (undo) action is triggered on button up
on the pen.  This should be fixed, but that will require better handling
of multiple clicks.

Some notes from development:
- The distance the pen is from the device doesn't seem to matter; it's
  just the 0x140 key event that the device is looking for.
- The timestamps on the events don't seem to matter.  Trying to fake
  a time doesn't seem to help or hurt.
- Adding an fsync after writing events doesn't seem to help.