Add machinery to detect different numbers of clicks
If another click starts within PRESS_TIMEOUT seconds of a previous
one, it is judged to be part of a multi-click sequence.  Only once
the timeout times out can we process a click event.

Right now, it should be set up to trigger undo and redo, but these
are not actually working.  I'm not sure why this is; triggering on
a button release does work (after the first one).

Also, delete some of the unused functionality.