Much longer delay before simulating multitouch
10 ms worked when we triggered on button release.  But with the
current system, nothing gets registered if we pause for 500 ms or
less.  Not sure why this is different.

Unfortunately, this results in a 1 s delay from the command to the
action: 0.5 s for the multiple-click detection to decide there's
no other clicks coming, and then 0.5 s to pause after shutting off
pen events.  This is too long to feel reactive.  Perhaps we can
grab the pen device in anticipation of an upcoming multitouch event?