Get toggle mode working
This is done more directly in the main loop, rather than holding
state in the separate toggleMode function.  Remarkable resets the
eraser mode when the pen moves away (as shown by a BTN_TOOL_PEN
event), so in the toggle eraser mode we have to look for these
events and trigger the eraser at the same time.

FWIW, clicking the button triggers a BTN_TOOL_PEN 0 and then a
BTN_TOOL_PEN 1 event.  Not sure why.