Clean up old code
The pressMode function is basically a single event call now. Switch
so that the writeEventVals handles setting the time, and get rid of
the old writeEvent function.  Get rid of unused arguments in the
main function and normalize the indentation.  We only need two
constants from the header file, so copy them into the main file
and get rid of the header file.