Reduce press timeout, to improve feel
Mainly, we cut the time in half, and measure since the most recent
press or release.  That way we can get a quicker timeout if the
press->release time was really short, while always allowing some
time after release for another press.

Also, we shrunk the new time a bit.  It still feels good.