Decrease delay in simulating multitouch
As soon as we get into a place where multitouch is likely, we grab
the pen device so that the device's timeout can begin.  Then when
it's time to write the multitouch events, we wait to ensure enough
time has transpired since the grab.  In practice, that seems to be
about 220 ms for double clicks, and 0 for triple clicks, which is
noticeably better than the 550 ms we were waiting before.