Trigger undo/redo through keyboard shortcuts
With uinput, we can create a fake keyboard on the device.  Ctrl-Z
and Ctrl-Shift-Z combinations sent through this trigger undo and
redo, even when the soft keyboard isn't up.

Oddly enough, the undo action works whether or not the pen is near
the surface.  But the redo action only works when the pen isn't
near.  (Wat?)  So we leave the pen grabbing code in place.

The Remarkable docs suggest that the device might be in "Mac" mode,
in which case these use Alt and not control.  So we probably should
add a switch to allow that.